Ad's mapper

+++ src/main/java/ru/skypro/homework/mapper/UserMapper.java
@@ -18,14 +18,14 @@ public class UserMapper {
 
     public UserDTO userEntityToUserDTO(UserEntity userEntity) {
         UserDTO userDTO = mapper.getMapper().map(userEntity, UserDTO.class);
-        userDTO.setId(Math.toIntExact(userEntity.getUserId()));
+        userDTO.setId(userEntity.getUserId().intValue());
         userDTO.setImage("нужно подправить маппер пока нет реализации с image");
         return userDTO;
     }
 
     public UserEntity userDTOToUserEntity(UserDTO userDTO) {
         UserEntity userEntity = mapper.getMapper().map(userDTO, UserEntity.class);
-        userEntity.setUserId(Long.valueOf(userDTO.getId()));
+        userEntity.setUserId(userDTO.getId().longValue());
         return userEntity;
     }
 }
